requirements.txt
pandas
streamlit
plotly
python-dotenv
joblib
Pillow
